pyproject.toml
[tool.poetry]
name = "quevedo"
version = "0.1.0"
description = ""
authors = ["Antonio F. G. Sevilla <afgs@ucm.es>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.1"
pillow = "^7.1.1"
pyyaml = "^5.3.1"
flask = { version = "^1.1.2", optional = true }
forcelayout = { version = "^1.0.6", optional = true }

[tool.poetry.extras]
tagger = [ "flask" ]
force_layout = [ "forcelayout" ]

[tool.poetry.scripts]
quevedo = "quevedo.cli:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
